finalized script
Script for project finalized. We used linear regression, polynomial
transformation, box cox transformation and logarithmic transformation.
Models were created for 100M, 200M, 1500M long Jump and shot put for
both Men and women events. Each model was selected by adj R^2 and
plotted to see how well it fits. Last, we used each of the models to
predict the future events to see if any of the models will converge to
some limit or be realistic